Minor fixes

+++ includes/dbActions.php
@@ -3,22 +3,28 @@ include 'dbHandler.php';
 
 //If post method has been submitted with name set to 'dbadd'
 if(isset($_POST['dbadd'])){
-    $email = $_POST['dbadd'];
-    //Checks if post contains anything
-    if($email != ""){
-        //Query for inserting new email to table, $email is value got from the post
-        //Email is set to unique in table, so there can not be duplicates
-        $query = "INSERT INTO submits (email) values ('$email')";
-        mysqli_query($GLOBALS['conn'], $query);       
+    //Checks if there is connection to MySQL
+    if($GLOBALS['conn']){
+        $email = $_POST['dbadd'];
+        //Checks if post contains anything
+        if($email != ""){
+            //Query for inserting new email to table, $email is value got from the post
+            //Email is set to unique in table, so there can not be duplicates
+            $query = "INSERT INTO submits (email) values ('$email')";
+            mysqli_query($GLOBALS['conn'], $query);       
+        }
     }
     //Directs browser to index.php
     header("Location: ../index.php");    
 }
 
 //If post method has been submitted with name set to 'erasedb'
-if(isset($_POST['erasedb'])){;
-    //Query for deleting the mysql table and creating a new empty one
-    mysqli_query($GLOBALS['conn'], "TRUNCATE TABLE submits;");
+if(isset($_POST['erasedb'])){
+    //Checks if there is connection to MySQL
+    if($GLOBALS['conn']){    
+        //Query for deleting the mysql table and creating a new empty one
+        mysqli_query($GLOBALS['conn'], "TRUNCATE TABLE submits;");
+    }
     //Directs browser to index.php
     header("Location: ../index.php");    
 }
